fix: safe_z_home virtual endstop compatibility
When combining a dockable probe with the 'safe_z_home' module, an error
could be raised after homing (e.g. "Must home axis first"). This was a
result of forcibly "unhoming" as part of dockable_probe's z-hop. The
unique combination of dockable probe, virtual Z endstop, safe_z_home,
and z-hop was failing to revalidate the homed axes before performing the
z-hop. (The use of '_last_homed' is a holdover from commit b6cc2f5.)