output_pin: Improve GCodeRequestQueue timing on duplicate requests
The GCodeRequestQueue code handles duplicate requests by detecting a
repeat of the last command.  If a duplicate is found, no scheduling is
performed.  However, in this case, the flush_notification() code
should continue looking for requests to schedule.

Update the GCodeRequestQueue code to use "discard" and "delay"
actions to indicate what steps the _flush_notification() callback
should take.  This simplifies the code and allows for future uses.

Signed-off-by: Kevin O'Connor <kevin@koconnor.net>