Warnings added to 5160
GS out of bounds warnings added to 5160.

Target current won't be able to be achieved if GS is > 256. Default set GS to 256 if it was over. New code pushes a warning to the user about the CS value being too low and needing to be higher to achieve the proper current.

TODO: better ihold handling.