neopixel: Support using SPI hardware to send data
This change allows the use of a SPI bus to send correctly timed data to
a string of neopixels. The idea is to take advantage of hardware to do
the timing; this does not magically make these SPI devices. As such, it
is only allowed on platforms that support hardware SPI.

The main motivation is that the bitbang approach cannot be used directly
from a Raspberry Pi secondary MCU as the bitbang timing is not precise
enough. Using a SPI bus to send data to neopixels is a commonly used
technique among other open source projects, and seems to work well in
Kalico, too.

On the devices I have tested (Raspberry Pi 5) the available SPI speeds
work with the data patterns that I have hardcoded, but it is conceivable
that it may be necessary to make the data configurable or automatically
generated based on requested bus speed in order to support some
neopixel-like devices.

In order to support Neopixel strings longer than 8-10 pixels (depending
on the number of colours), this also adds a `spidev_transfer_large` to
the linux platform which uses a `size_t` data length instead of
`uint8_t`.

Signed-off-by: Russell Cloran <rcloran@gmail.com>