Update README.md
Added additional instructions for KIAUH v4 and KIAUH v6 as v6 is not in full release yet. Due to changes in the KIAUH rewrite to v6, if you add the repos to your klipper_repos.txt copying from the latest readme update for using v4. You add an extra space between the repo link and branch to the .kiauh.ini after selecting the new repo and KIAUH will fail to load until that is corrected.