Replace auto-generated description with param help
This adds "params" to printer.status.gcode.commands. The basic structure
is something like

{
    "SCRUB_NOZZLE": {
        "help": "Scrub the nozzle",
        "params": {
            "COUNT": { "type": "int", "default": 5 },
            "SPEED": { "type": "float", "default": 100 },
        }
    }
}

The "type" is the name of the annotation
"required" is set when a parameter has no default
"default" is set when a non-null default value is provided.
  Note: this must be json serializable
"choices" when the parameter is an enum, this is an array of the enum values

This should give frontends plenty of information to make much nicer
interfaces for calling macros